LayoutTests/http/tests/security/xssAuditor/script-tag-inside-svg-tag-expected.txt

CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

 Ensures HTML entities are recognized in script blocks in a context where CDATA is allowed.
